requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0